refactor: reorganize controller methods for better separation of concerns
- Move prescription methods (getPrescriptionDetails, deletePrescription) to PrescriptionController
- Move product methods (addProduct, deleteProduct) to ProductController
- Move user methods (addUser, addUserPost, deleteUsers, verifyPassword) to UserController
- Move profile update methods (updatePassword, updateProfile) to ProfileController
- Update routes to reference appropriate controllers
- Keep profile display method in DashboardController for dashboard functionality

This improves code organization by grouping related functionality into dedicated controllers.
