+++ app/Http/Controllers/ProfileController.php
@@ -1,0 +1,117 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
+
+class ProfileController extends Controller
+{
+    /**
+     * Update user password
+     */
+    public function updatePassword(Request $request)
+    {
+        try {
+            $request->validate([
+                'current_password' => 'required',
+                'new_password' => 'required|min:8',
+                'new_password_confirmation' => 'required|same:new_password',
+            ]);
+
+            $user = auth()->user();
+            
+            // Check if current password is correct
+            if (!Hash::check($request->current_password, $user->password)) {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'Current password is incorrect.'
+                ], 422);
+            }
+
+            // Update the password
+            $user->update([
+                'password' => Hash::make($request->new_password)
+            ]);
+
+            return response()->json([
+                'success' => true,
+                'message' => 'Password updated successfully!'
+            ]);
+            
+        } catch (\Illuminate\Validation\ValidationException $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation failed.',
+                'errors' => $e->errors()
+            ], 422);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'An error occurred while updating password.'
+            ], 500);
+        }
+    }
+
+    /**
+     * Update user profile
+     */
+    public function updateProfile(Request $request)
+    {
+        try {
+            // Enhanced validation with custom messages
+            $request->validate([
+                'name' => 'required|string|min:2|max:255',
+                'email' => 'required|email|unique:users,email,' . auth()->id(),
+                'role' => 'required|string|in:admin,user',
+            ], [
+                'name.required' => 'Name is required.',
+                'name.min' => 'Name must be at least 2 characters long.',
+                'name.max' => 'Name must not exceed 255 characters.',
+                'email.required' => 'Email is required.',
+                'email.email' => 'Please enter a valid email address.',
+                'email.unique' => 'This email is already taken by another user.',
+                'role.required' => 'User role is required.',
+                'role.in' => 'Invalid user role selected.',
+            ]);
+
+            $user = auth()->user();
+            
+            // Update the user data
+            $updated = $user->update([
+                'name' => trim($request->name),
+                'email' => trim($request->email),
+                'role' => $request->role,
+            ]);
+
+            if (!$updated) {
+                throw new \Exception('Failed to update profile in database.');
+            }
+
+            // Refresh the user model to get updated data
+            $user->refresh();
+
+            return response()->json([
+                'success' => true,
+                'message' => 'Profile updated successfully!',
+                'user' => [
+                    'name' => $user->name,
+                    'email' => $user->email,
+                    'role' => $user->role,
+                ]
+            ]);
+            
+        } catch (\Illuminate\Validation\ValidationException $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Validation failed.',
+                'errors' => $e->errors()
+            ], 422);
+        } catch (\Exception $e) {
+            return response()->json([
+                'success' => false,
+                'message' => 'An error occurred while updating profile: ' . $e->getMessage()
+            ], 500);
+        }
+    }
+}
